# book-network/src/main/java/com/pro/book/auth/AuthenticationController.java
package com.pro.book.auth;

import io.swagger.v3.oas.annotations.tags.Tag;
import jakarta.mail.MessagingException;
import jakarta.validation.Valid;
import lombok.RequiredArgsConstructor;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.*;

@RestController
@RequestMapping("auth")
@RequiredArgsConstructor
@Tag(name = "Authentication")
public class AuthenticationController {

    private final AuthenticatonService authenticatonService;

    @PostMapping("/register")
    @ResponseStatus(HttpStatus.ACCEPTED)
    public ResponseEntity<?> register( @RequestBody @Valid ResgistrationRequest request )
            throws MessagingException {
        authenticatonService.register(request);
        return ResponseEntity.accepted().build();
    }

    @PostMapping("/authenticate")
    public ResponseEntity<AuthenticationResponse> authenticate(@RequestBody @Valid AuthenticationRequest request){
        return ResponseEntity.ok(authenticatonService.authenticate(request));
    }

    @GetMapping("/activate-account")
    public void confirm( @RequestParam String token) throws MessagingException {
        authenticatonService.activateAccount(token);
    }

}